requirements.txt
asgiref==3.7.2
backoff==2.2.1
Deprecated==1.2.14
Django==4.2.7
googleapis-common-protos==1.61.0
grpcio==1.59.2
importlib-metadata==6.8.0
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-django==0.41b0
opentelemetry-instrumentation-wsgi==0.41b0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
protobuf==4.25.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
wrapt==1.15.0
zipp==3.17.0
